requirements.txt
appdirs==1.4.3
coverage==4.3.4
Django==1.11
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
